+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- New Option `--hidden` to hide the console window on Windows (useful for GUI applications)
 
 ## [v1.2.0] - 2026-03-07
 ### Added